Clean up: reorganize namespaces
