+++ src/TypingMachine/CodeLoading/ICodeBase.cs
@@ -1,5 +1,5 @@
 ﻿using System.Collections.Generic;
-using TypingMachine.Entities;
+using TypingMachine.Domain;
 
 namespace TypingMachine.CodeLoading
 {
