+++ src/TypingMachine/CodeParsing/Finders/AccessModifierFinder.cs
@@ -1,8 +1,8 @@
 ﻿using System.Linq;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
-using TypingMachine.Entities;
+using TypingMachine.Domain;
 
-namespace TypingMachine.CodeFinders
+namespace TypingMachine.CodeParsing.Finders
 {
     public class AccessModifierFinder
     {
