+++ src/TypingMachine/CodeParsing/Finders/NamespaceFinder.cs
@@ -2,9 +2,9 @@
 using System.Linq;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
-using TypingMachine.Entities;
+using TypingMachine.Domain;
 
-namespace TypingMachine.CodeFinders
+namespace TypingMachine.CodeParsing.Finders
 {
     public class NamespaceFinder
     {
