+++ src/TypingMachine/Domain/Builders/FieldBuilder.cs
@@ -1,6 +1,4 @@
-﻿using TypingMachine.Entities;
-
-namespace TypingMachine.Builders
+﻿namespace TypingMachine.Domain.Builders
 {
     public class FieldBuilder
     {
