+++ src/TypingMachine/Domain/InterfaceEntity.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using TypingMachine.Abstractions;
 
-namespace TypingMachine.Entities
+namespace TypingMachine.Domain
 {
     public class InterfaceEntity : TypeEntity
     {
