+++ tests/TypingMachine.Tests/CodeFinding/NamespaceFinderTests/FindingNamespaceTestContexts.cs
@@ -1,4 +1,4 @@
-﻿using TypingMachine.Entities;
+﻿using TypingMachine.Domain;
 
 namespace TypingMachine.Tests.CodeFinding.NamespaceFinderTests
 {
