+++ tests/TypingMachine.Tests/CodeWalking/TypeWalkerTests/WhenDiscoveringClassEntity.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using FluentAssertions;
 using Microsoft.CodeAnalysis.CSharp;
-using TypingMachine.CodeWalkers;
+using TypingMachine.CodeParsing.Walkers;
 using Xunit;
 
 namespace TypingMachine.Tests.CodeWalking.TypeWalkerTests
