+++ tests/TypingMachine.Tests/Entities/NamespaceIdentifierTests/WhenCreatingNamespaceIdentifier.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using FluentAssertions;
-using TypingMachine.Entities;
+using TypingMachine.Domain;
 using Xunit;
 
 namespace TypingMachine.Tests.Entities.NamespaceIdentifierTests
