+++ tests/TypingMachine.Tests/Entities/NamespaceIdentifierTests/WhenReadingFullName.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 using FluentAssertions;
-using TypingMachine.Entities;
+using TypingMachine.Domain;
 using Xunit;
 
 namespace TypingMachine.Tests.Entities.NamespaceIdentifierTests
