+++ tests/TypingMachine.Tests/Entities/TypeEntities/WhenCreatingInterfaceEntity.cs
@@ -1,8 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
 using FluentAssertions;
-using TypingMachine.Builders;
-using TypingMachine.Entities;
+using TypingMachine.Domain;
+using TypingMachine.Domain.Builders;
 using TypingMachine.Tests.Utilities;
 using Xunit;
