+++ tests/TypingMachine.Tests/Entities/TypeIdentifierTests/WhenComparingTypeIdentifiers.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 using FluentAssertions;
-using TypingMachine.Entities;
+using TypingMachine.Domain;
 using TypingMachine.Tests.Utilities;
 using Xunit;
