+++ tests/TypingMachine.Tests/Entities/UsingEntityTests/WhenCreatingUsingEntity.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 using FluentAssertions;
-using TypingMachine.Entities;
+using TypingMachine.Domain;
 using Xunit;
 
 namespace TypingMachine.Tests.Entities.UsingEntityTests
