+++ tests/TypingMachine.Tests/TypeVisitorTests/WhenVisitingTypeEntity.cs
@@ -1,8 +1,8 @@
 ﻿using System.Collections.Generic;
 using Moq;
 using TypingMachine.Abstractions;
-using TypingMachine.Builders;
-using TypingMachine.Entities;
+using TypingMachine.Domain;
+using TypingMachine.Domain.Builders;
 using TypingMachine.Tests.Utilities;
 using Xunit;
