+++ tests/TypingMachine.Tests/Utilities/NamespaceIdentifierExtensions.cs
@@ -1,4 +1,4 @@
-﻿using TypingMachine.Entities;
+﻿using TypingMachine.Domain;
 
 namespace TypingMachine.Tests.Utilities
 {
